Rename EditAdditionalRepositoriesButton to AddAdditionalRepositoriesButton
